codes/dprojx/requirements.txt
Django==2.2.16
django-phonenumber-field
phonenumbers
django-ses
djangorestframework
boto3
requests
django-crispy-forms
Pillow
psycopg2-binary
gunicorn
newrelic
django-allauth
django-rest-auth
django-cors-headers
stripe
django-magic-link
drf_yasg
django-dbbackup
slack
pandas
slackclient==2.1.0
